third_party/WebKit/LayoutTests/platform/mac/fast/forms/select/popup-menu-ax-expected.txt

CONSOLE ERROR: line 44: Uncaught TypeError: Cannot read property 'addEventListener' of null
Tests if typing an arrow key dispatches a |MenuListItemSelected| a11y event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Unexpected error: Uncaught TypeError: Cannot read property 'addEventListener' of null
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
